+++ examples/vanilla/a.js
@@ -1,6 +1,9 @@
+/* eslint-disable no-eval */
+/* eslint-disable no-new-func */
 export function test() {
   const el = document.createElement('div')
   el.innerText = 'vanilla'
   document.body.append(el)
-  const c = eval('console.log(1)')
+  eval('console.log(document.body)')
+  new Function('const d=10')('')
 }