+++ packages/helper/src/index.ts
@@ -4,9 +4,11 @@ export { Merak, Props, NameSpace }
 
 export type MerakEvent = 'mount' | 'destroy' | 'hidden' | 'unmount' | 'relunch' | 'show'
 
+export const isMerak = !!window.isMerak
+
 // get real window
 export function $window(): Window {
-  return isMerak() ? window.rawWindow : window
+  return isMerak ? window.rawWindow : window
 }
 
 export function $document(): Document {
@@ -28,7 +30,7 @@ export function getInstance(id: string) {
  * @no_recommand 不建议使用
  */
 export function $jump(project: string, to: string) {
-  if (!isMerak())
+  if (!isMerak)
     return
   const instance = getInstance(project)
   if (instance) {
@@ -42,12 +44,9 @@ export function $eventName(event: string) {
   return `merak_${event}`
 }
 
-export function isMerak() {
-  return !!window.isMerak
-}
 // sub app baseUrl
 export function $base() {
-  return isMerak() ? $instance()!.url : location.origin
+  return isMerak ? $instance()!.url : location.origin
 }
 
 export function $body(): HTMLElement {
@@ -58,82 +57,58 @@ export function $head(): HTMLHeadElement {
   return $document().head
 }
 
-export function $on(eventName: MerakEvent, cb: () => any): () => void {
+export function $on(eventName: MerakEvent, cb: (flag?: string) => any): () => void {
   const event = $eventName(eventName)
-  if (isMerak()) {
-    window.addEventListener(event, cb)
-    return () => window.removeEventListener(event, cb)
+  if (isMerak) {
+    const callback = function (e: any) {
+      cb(e.detail)
+    }
+    window.addEventListener(event, callback)
+    return () => window.removeEventListener(event, callback)
   }
   return () => { }
 }
 
-export function $once(eventName: MerakEvent, cb: () => any): () => void {
+export function $once(eventName: MerakEvent, cb: (flag?: string) => any): () => void {
   const event = $eventName(eventName)
-  if (isMerak()) {
-    window.addEventListener(event, cb, { once: true })
-    return () => window.removeEventListener(event, cb)
+  if (isMerak) {
+    const callback = function (e: any) {
+      cb(e.detail)
+    }
+    window.addEventListener(event, callback, { once: true })
+    return () => window.removeEventListener(event, callback)
   }
 
   return () => { }
-}
-
-// I don't sure if it is important
-export function $onShow(cb: () => any) {
-  return $on('show', cb)
-}
-export function $onRelunch(cb: () => any) {
-  return $on('relunch', cb)
-}
-
-export function $onHidden(cb: () => any) {
-  return $on('hidden', cb)
-}
-
-export function $onDestroy(cb: () => any) {
-  return $on('destroy', cb)
 }
 
 /**
  * $onMount $onUnmount $onExec run both in merak and individual app
  */
 
-export function $onMount(cb: () => any) {
-  return isMerak() ? $on('mount', cb) : cb()
+export function $onMount(cb: (flag?: string) => any) {
+  return isMerak ? $on('mount', cb) : cb()
 }
 
-export function $onUnmount(cb: () => any) {
-  if (isMerak()) {
+export function $onUnmount(cb: (flag?: 'destroy' | string) => any) {
+  if (isMerak) {
     const fn = $on('unmount', cb)
     return fn
   }
   else {
-    window.addEventListener('unload', cb)
+    const callback = function (e: any) {
+      cb(e.detail)
+    }
+    window.addEventListener('beforeunload', callback)
     return () => {
-      window.removeEventListener('unload', cb)
+      window.removeEventListener('beforeunload', callback)
     }
   }
 }
 
-export function $onExec(cb: () => any) {
-  cb()
-  if (isMerak())
-    $on('relunch', cb)
-}
-
-// export function $onExec(cb: () => any) {
-//   if (!isMerak())
-//     cb()
-//   const fn1 = $once('mount', cb)
-//   const fn2 = $on('relunch', cb)
-//   return () => {
-//     fn1()
-//     fn2()
-//   }
-// }
-
 // work for eval
 export function $sandbox(script: string) {
-  if (isMerak()) {
+  if (isMerak) {
     const fakeGlobalVar = window.$Merak.fakeGlobalVar
     return `with(window.$MerakMap.get(${fakeGlobalVar}).proxy){${script}}`
   }
@@ -142,9 +117,9 @@ export function $sandbox(script: string) {
 }
 // work for esm script
 export function $esm(script: string) {
-  if (isMerak()) {
+  if (isMerak) {
     const { fakeGlobalVar, nativeVars, customVars } = window.$Merak
-    return `const {${nativeVars.reduce((p: string, c: string) => `${p},${c}`)}}=${fakeGlobalVar};${createCustomVarProxy(fakeGlobalVar, customVars)}${script}`
+    return `const {${nativeVars.reduce((p: string, c: string) => `${p},${c}`)}}=${fakeGlobalVar};${createCustomVarProxy(fakeGlobalVar, customVars)}\n${script}`
   }
   return script
 }
@@ -173,8 +148,11 @@ export function $perf() {
   return $instance()?.perf
 }
 
+/**
+ * @experiment
+ */
 export function $stopBubble(isPass = false) {
-  if (isMerak()) {
+  if (isMerak) {
     ['click', 'mousemove', 'mousedown', 'mouseup', 'keydown', 'keyup'].forEach((eventName) => {
       document.body.addEventListener(eventName, (event) => {
         event.stopPropagation()
@@ -186,8 +164,14 @@ export function $stopBubble(isPass = false) {
 }
 /**
  * Tell the host application that it's time to uninstall,
- * only invoke inside $onDestroy
  */
-export function $done() {
+export function $deactive() {
   $instance()?.deactive()
+}
+
+/**
+ * @danger
+ */
+export function $destroy() {
+  $instance()?.destroy()
 }