+++ packages/helper/CHANGELOG.md
@@ -1,5 +1,17 @@
 # merak-helper
 
+## 2.0.0-alpha.1
+
+### Patch Changes
+
+- 70316cb: export isMerak
+
+## 2.0.0-alpha.0
+
+### Major Changes
+
+- 81c8105: break change: only leave mount and unmount event ,replace keep-alive with flag
+
 ## 1.0.1
 
 ### Patch Changes